+++ src/test/converter2/behavior/documentTag.ts
@@ -1,0 +1,5 @@
+/**
+ * @document ./document.md
+ */
+export interface HasDescriptor {
+}